requirements.txt
tensorflow-datasets
kaggle
matplotlib
scikit-learn
kagglehub
numpy
matplotlib